Move everything `PrimitiveSealed` into separate file
